# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pet Spa Booking API Documentation",
    "description": "RESTful API cho hệ thống đặt lịch Pet Spa",
    "version": "1.0.0",
    "contact": {
      "name": "Pet Spa Support",
      "email": "admin@petspa.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Pet Spa Booking API Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer token"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "userID": {"type": "integer", "example": 1},
          "name": {"type": "string", "example": "Nguyễn Văn A"},
          "email": {"type": "string", "example": "user@example.com"},
          "phone": {"type": "string", "example": "0123456789"},
          "address": {"type": "string", "example": "123 Đường ABC"},
          "role": {"type": "string", "example": "user"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Đăng ký tài khoản mới",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password", "password_confirmation"],
                "properties": {
                  "name": {"type": "string", "example": "Nguyễn Văn A"},
                  "email": {"type": "string", "format": "email", "example": "user@example.com"},
                  "password": {"type": "string", "format": "password", "example": "password123"},
                  "password_confirmation": {"type": "string", "format": "password", "example": "password123"},
                  "phone": {"type": "string", "example": "0123456789"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đăng ký thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "message": {"type": "string", "example": "Đăng ký thành công!"},
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {"$ref": "#/components/schemas/User"},
                        "access_token": {"type": "string", "example": "1|abc123..."},
                        "token_type": {"type": "string", "example": "Bearer"}
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {"description": "Validation Error"}
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Đăng nhập",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "format": "email", "example": "user@example.com"},
                  "password": {"type": "string", "format": "password", "example": "password123"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đăng nhập thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "message": {"type": "string", "example": "Đăng nhập thành công!"},
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {"$ref": "#/components/schemas/User"},
                        "access_token": {"type": "string"},
                        "token_type": {"type": "string", "example": "Bearer"}
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {"description": "Validation Error"}
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Đăng xuất",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Đăng xuất thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "message": {"type": "string", "example": "Đăng xuất thành công!"}
                  }
                }
              }
            }
          },
          "401": {"description": "Unauthorized"}
        }
      }
    },
    "/auth/user": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Lấy thông tin user hiện tại",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {"$ref": "#/components/schemas/User"}
                  }
                }
              }
            }
          },
          "401": {"description": "Unauthorized"}
        }
      }
    },
...
    "/auth/google/callback": {
      "post": {
        "tags": ["Google Auth"],
        "summary": "Xử lý callback từ Google",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code"],
                "properties": {
                  "code": {"type": "string", "description": "Authorization code từ Google"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Đăng nhập thành công"},
          "401": {"description": "Đăng nhập Google thất bại"}
        }
      }
    },
...
    "/home/search": {
      "get": {
        "tags": ["Home"],
        "summary": "Tìm kiếm dịch vụ",
        "parameters": [
          {"name": "q", "in": "query", "description": "Từ khóa tìm kiếm", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Thành công"}
        }
      }
    },
...
  }
}